requirements.txt
azure-ai-ml==1.12.0
azure-identity==1.15.0
azureml-mlflow==1.53.0
mlflow==2.9.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
pyyaml==6.0.1
joblib==1.3.2
pyarrow==14.0.1
delta-spark==2.4.0
pyspark==3.4.1
opencensus-ext-azure==1.1.9
azure-monitor-opentelemetry==1.2.0
scipy==1.11.3
pytest==7.4.3
pytest-cov==4.1.0
